添加字典相关功能

+++ src/main/java/com/ls/erp/controller/UserController.java
@@ -1,13 +1,14 @@
 package com.ls.erp.controller;
 
 import com.alibaba.fastjson.JSONObject;
+import com.ls.erp.annotation.PassAuthToken;
+import com.ls.erp.annotation.PassReturnToken;
 import com.ls.erp.entity.ResultInfo;
 import com.ls.erp.entity.RoleInfo;
-import com.ls.erp.entity.RoleUserInfo;
 import com.ls.erp.entity.UserInfo;
-import com.ls.erp.service.RoleService;
 import com.ls.erp.service.UserService;
 import com.ls.erp.utils.LogUtil;
+import com.ls.erp.utils.TokenUtil;
 import com.mysql.cj.core.util.StringUtils;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -17,7 +18,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
-import java.util.Iterator;
+import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 import java.util.Map;
 
@@ -28,12 +29,65 @@ public class UserController {
     @Resource
     LogUtil logUtil;
 
+    @Resource
+    TokenUtil tokenUtil;
+
     @Autowired
     private UserService userService;
 
-    @Autowired
-    private RoleService roleService;
     // 用户登陆
+    @PassAuthToken
+    @PostMapping("/user/login")
+    @ApiOperation("登陆用户")
+    @ApiResponses({
+            @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
+            @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
+    })
+    public ResultInfo loginUser(UserInfo userInfo, HttpServletResponse response) {
+        logUtil.in("请求登陆用户");
+        if (userInfo == null || StringUtils.isNullOrEmpty(userInfo.getLoginName()) || StringUtils.isNullOrEmpty(userInfo.getPassword())) {
+            logUtil.out("请求登陆用户 接收参数为空");
+            return ResultInfo.error("请求参数不正确");
+        }
+        UserInfo u = userService.findUserByLoginNameAndPassword(userInfo.getLoginName(), userInfo.getPassword());
+        if (u == null) {
+            logUtil.in("请求登陆用户失败 未找到该用户");
+            return ResultInfo.error("未找到该用户");
+        }
+
+        logUtil.in("请求登陆用户成功");
+        response.setHeader("token", tokenUtil.getToken(u));
+        return ResultInfo.success("登陆成功");
+    }
+
+    // 用户注册
+    @PassAuthToken
+    @PassReturnToken
+    @PostMapping("/user/sign")
+    @ApiOperation("注册用户")
+    @ApiResponses({
+            @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
+            @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
+    })
+    public ResultInfo signUser(UserInfo userInfo) {
+        logUtil.in("请求注册用户");
+        if (userInfo == null || StringUtils.isNullOrEmpty(userInfo.getLoginName()) || StringUtils.isNullOrEmpty(userInfo.getPassword()) || StringUtils.isNullOrEmpty(userInfo.getName())) {
+            logUtil.out("请求注册用户 接收参数为空");
+            return ResultInfo.error("请求参数不正确");
+        }
+
+        if (userService.existsLoginName(userInfo.getLoginName())) {
+            logUtil.out("请求注册用户 用户已存在");
+            return ResultInfo.error("用户已存在");
+        }
+
+        userService.addUser(userInfo);
+
+        logUtil.in("请求注册用户成功");
+        return ResultInfo.success("注册成功");
+    }
 
     // 新增用户
     @PostMapping("/user")
@@ -41,14 +95,20 @@ public class UserController {
     @ApiResponses({
             @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
             @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
     })
     public ResultInfo addUser(UserInfo userInfo) {
         logUtil.in("请求添加用户");
 
         if (userInfo == null || StringUtils.isNullOrEmpty(userInfo.getLoginName()) || StringUtils.isNullOrEmpty(userInfo.getPassword()) || StringUtils.isNullOrEmpty(userInfo.getName())) {
-            logUtil.out("请求给角色添加权限 接收参数为空");
+            logUtil.out("请求添加用户 接收参数为空");
             return ResultInfo.error("请求参数不正确");
         }
+        if (userService.existsLoginName(userInfo.getLoginName())) {
+            logUtil.out("请求注册用户 用户已存在");
+            return ResultInfo.error("用户已存在");
+        }
+
         userService.addUser(userInfo);
 
         logUtil.out("请求添加用户成功");
@@ -61,6 +121,7 @@ public class UserController {
     @ApiResponses({
             @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
             @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
     })
     public ResultInfo deleteUser(UserInfo userInfo) {
         logUtil.in("请求删除用户");
@@ -69,6 +130,11 @@ public class UserController {
             logUtil.out("请求删除角色 接收参数为空");
             return ResultInfo.error("请求参数不正确");
         }
+        if (!userService.existsById(userInfo.getId())) {
+            logUtil.out("请求注册用户 用户不存在");
+            return ResultInfo.error("用户不存在");
+        }
+
         userService.deleteUser(userInfo);
 
         logUtil.out("请求删除用户成功");
@@ -81,6 +147,7 @@ public class UserController {
     @ApiResponses({
             @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
             @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
     })
     public ResultInfo updateUser(UserInfo userInfo) {
         logUtil.in("请求编辑用户");
@@ -105,6 +172,7 @@ public class UserController {
     @ApiResponses({
             @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
             @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
     })
     public ResultInfo findAllUser() {
         logUtil.in("请求查询所有用户");
@@ -113,12 +181,37 @@ public class UserController {
         return ResultInfo.success(userInfoList, "查询成功");
     }
 
+    @GetMapping("/user/find/{id}")
+    @ApiOperation("根据id查看用户")
+    @ApiResponses({
+            @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
+            @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
+    })
+    public ResultInfo findUserById(@PathVariable("id")int id) {
+        logUtil.in("请求查询ID用户");
+        if (id == 0) {
+            // id 错误
+            logUtil.out("请求查询ID用户 接收参数为空");
+            return ResultInfo.error("请求参数不正确");
+        }
+        UserInfo userInfo = userService.findUserById(id);
+        if (userInfo == null) {
+            logUtil.out("请求查询ID用户 未找到");
+            return ResultInfo.error("未找到");
+        }
+
+        logUtil.out("请求查询ID用户 成功");
+        return ResultInfo.success(userInfo, "查找成功");
+    }
+
     // 查询相关用户的角色
     @GetMapping("/user/role/{id}")
     @ApiOperation("查看用户角色")
     @ApiResponses({
             @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
             @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
     })
     public ResultInfo listRole(@PathVariable("id")int id) {
         logUtil.in("请求给用户添加角色");
@@ -140,6 +233,7 @@ public class UserController {
     @ApiResponses({
             @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
             @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
     })
     public ResultInfo addRole(@RequestBody Map<String, Object> req) {
         logUtil.in("请求给用户添加角色");
@@ -167,6 +261,7 @@ public class UserController {
     @ApiResponses({
             @ApiResponse(code = 0, message = "请求成功", response = ResultInfo.class),
             @ApiResponse(code = -1, message = "请求失败", response = ResultInfo.class),
+            @ApiResponse(code = -2, message = "token失效", response = ResultInfo.class),
     })
     public ResultInfo deleteRole(@RequestBody Map<String, Object> req) {
         logUtil.in("请求给用户删除角色");
